+++ backend/core/tests/fixture_chunks.py
@@ -5,7 +5,6 @@ from uuid import uuid4
 from langchain_core.embeddings import DeterministicFakeEmbedding
 from langchain_core.messages.ai import AIMessageChunk
 from langchain_core.vectorstores import InMemoryVectorStore
-
 from quivr_core.chat import ChatHistory
 from quivr_core.config import LLMEndpointConfig, RetrievalConfig
 from quivr_core.llm import LLMEndpoint
@@ -19,7 +18,9 @@ async def main():
 
     llm = LLMEndpoint.from_config(retrieval_config.llm_config)
     chat_history = ChatHistory(uuid4(), uuid4())
-    rag_pipeline = QuivrQARAGLangGraph(retrieval_config=retrieval_config, llm=llm, vector_store=vec)
+    rag_pipeline = QuivrQARAGLangGraph(
+        retrieval_config=retrieval_config, llm=llm, vector_store=vec
+    )
 
     conversational_qa_chain = rag_pipeline.build_chain()
 
@@ -36,7 +37,10 @@ async def main():
             config={"metadata": {}},
         ):
             kind = event["event"]
-            if kind == "on_chat_model_stream" and event["metadata"]["langgraph_node"] == "generate":
+            if (
+                kind == "on_chat_model_stream"
+                and event["metadata"]["langgraph_node"] == "generate"
+            ):
                 chunk = event["data"]["chunk"]
                 dict_chunk = {
                     k: v.dict() if isinstance(v, AIMessageChunk) else v