+++ backend/worker/quivr_worker/celery_worker.py
@@ -8,12 +8,14 @@ from dotenv import load_dotenv
 from quivr_api.celery_config import celery
 from quivr_api.logger import get_logger
 from quivr_api.models.settings import settings
+from quivr_api.modules.assistant.repository.tasks import TasksRepository
+from quivr_api.modules.assistant.services.tasks_service import TasksService
 from quivr_api.modules.brain.integrations.Notion.Notion_connector import NotionConnector
 from quivr_api.modules.brain.repository.brains_vectors import BrainsVectors
 from quivr_api.modules.brain.service.brain_service import BrainService
 from quivr_api.modules.dependencies import get_supabase_client
 from quivr_api.modules.knowledge.repository.knowledges import KnowledgeRepository
-from quivr_api.modules.knowledge.repository.storage import Storage
+from quivr_api.modules.knowledge.repository.storage import SupabaseS3Storage
 from quivr_api.modules.knowledge.service.knowledge_service import KnowledgeService
 from quivr_api.modules.notification.service.notification_service import (
     NotificationService,
@@ -29,6 +31,7 @@ from sqlalchemy.ext.asyncio import AsyncEngine, create_async_engine
 from sqlmodel import Session, text
 from sqlmodel.ext.asyncio.session import AsyncSession
 
+from quivr_worker.assistants.assistants import process_assistant
 from quivr_worker.check_premium import check_is_premium
 from quivr_worker.process.process_s3_file import process_uploaded_file
 from quivr_worker.process.process_url import process_url_func
@@ -39,7 +42,7 @@ from quivr_worker.syncs.process_active_syncs import (
     process_sync,
 )
 from quivr_worker.syncs.store_notion import fetch_and_store_notion_files_async
-from quivr_worker.utils import _patch_json
+from quivr_worker.utils.utils import _patch_json
 
 load_dotenv()
 
@@ -58,7 +61,7 @@ sync_user_service = SyncUserService()
 sync_files_repo_service = SyncFilesRepository()
 brain_service = BrainService()
 brain_vectors = BrainsVectors()
-storage = Storage()
+storage = SupabaseS3Storage()
 notion_service: SyncNotionService | None = None
 async_engine: AsyncEngine | None = None
 engine: Engine | None = None
@@ -94,6 +97,63 @@ def init_worker(**kwargs):
 @celery.task(
     retries=3,
     default_retry_delay=1,
+    name="process_assistant_task",
+    autoretry_for=(Exception,),
+)
+def process_assistant_task(
+    assistant_id: str,
+    notification_uuid: str,
+    task_id: int,
+    user_id: str,
+):
+    logger.info(
+        f"process_assistant_task started for assistant_id={assistant_id}, notification_uuid={notification_uuid}, task_id={task_id}"
+    )
+    print("process_assistant_task")
+
+    loop = asyncio.get_event_loop()
+    loop.run_until_complete(
+        aprocess_assistant_task(
+            assistant_id,
+            notification_uuid,
+            task_id,
+            user_id,
+        )
+    )
+
+
+async def aprocess_assistant_task(
+    assistant_id: str,
+    notification_uuid: str,
+    task_id: int,
+    user_id: str,
+):
+    async with AsyncSession(async_engine) as async_session:
+        try:
+            await async_session.execute(
+                text("SET SESSION idle_in_transaction_session_timeout = '5min';")
+            )
+            tasks_repository = TasksRepository(async_session)
+            tasks_service = TasksService(tasks_repository)
+
+            await process_assistant(
+                assistant_id,
+                notification_uuid,
+                task_id,
+                tasks_service,
+                user_id,
+            )
+
+        except Exception as e:
+            await async_session.rollback()
+            raise e
+        finally:
+            await async_session.close()
+
+
+@celery.task(
+    retries=3,
+    default_retry_delay=1,
     name="process_file_task",
     autoretry_for=(Exception,),
     dont_autoretry_for=(FileExistsError,),
@@ -110,10 +170,6 @@ def process_file_task(
 ):
     if async_engine is None:
         init_worker()
-
-    logger.info(
-        f"Task process_file started for file_name={file_name}, knowledge_id={knowledge_id}, brain_id={brain_id}, notification_id={notification_id}"
-    )
 
     loop = asyncio.get_event_loop()
     loop.run_until_complete(
@@ -170,6 +226,8 @@ async def aprocess_file_task(
                     integration_link=source_link,
                     delete_file=delete_file,
                 )
+                session.commit()
+            await async_session.commit()
         except Exception as e:
             session.rollback()
             await async_session.rollback()
@@ -196,19 +254,29 @@ def process_crawl_task(
     )
     global engine
     assert engine
-    with Session(engine, expire_on_commit=False, autoflush=False) as session:
-        vector_repository = VectorRepository(session)
-        vector_service = VectorService(vector_repository)
-        loop = asyncio.get_event_loop()
-        loop.run_until_complete(
-            process_url_func(
-                url=crawl_website_url,
-                brain_id=brain_id,
-                knowledge_id=knowledge_id,
-                brain_service=brain_service,
-                vector_service=vector_service,
+    try:
+        with Session(engine, expire_on_commit=False, autoflush=False) as session:
+            session.execute(
+                text("SET SESSION idle_in_transaction_session_timeout = '5min';")
             )
-        )
+            vector_repository = VectorRepository(session)
+            vector_service = VectorService(vector_repository)
+            loop = asyncio.get_event_loop()
+            loop.run_until_complete(
+                process_url_func(
+                    url=crawl_website_url,
+                    brain_id=brain_id,
+                    knowledge_id=knowledge_id,
+                    brain_service=brain_service,
+                    vector_service=vector_service,
+                )
+            )
+            session.commit()
+    except Exception as e:
+        session.rollback()
+        raise e
+    finally:
+        session.close()
 
 
 @celery.task(name="NotionConnectorLoad")